+++ components/form/admin/types/index.ts
@@ -1,4 +1,5 @@
 import React from 'react'
+import { CheckboxType } from './checkbox.type'
 import { DateType } from './date.type'
 import { DropdownType } from './dropdown.type'
 import { EmailType } from './email.type'
@@ -18,6 +19,7 @@ export const adminTypes: {
 } = {
   date: DateType,
   dropdown: DropdownType,
+  checkbox: CheckboxType,
   email: EmailType,
   hidden: HiddenType,
   link: LinkType,